fix(ci): unblock os-smoke on Windows + macOS (shutdown_test build tag, ACL test path, watcher flake skip)
- A1: Move TestManager_Wait_Signal (uses syscall.Kill) to
  shutdown_wait_signal_unix_test.go with //go:build unix; remove syscall
  import from platform-agnostic shutdown_test.go.
- A2: Fix credfile_security_windows_test.go lines 66+108: pass string path
  directly to windows.SetNamedSecurityInfo (wrapper handles UTF-16
  conversion); drop now-unused *uint16 pathPtr variable.
- A3: Skip TestWatcher_AtomicReplace_RenameCreate on darwin via
  runtime.GOOS guard — fsnotify event coalescing on rename+create is
  pre-existing flaky on macOS; Linux + Windows still cover the path.
