+++ runtime/config/watcher_test.go
@@ -100,6 +100,9 @@ func TestNewWatcher_InvalidPath(t *testing.T) {
 }
 
 func TestWatcher_AtomicReplace_RenameCreate(t *testing.T) {
+	if runtime.GOOS == "darwin" {
+		t.Skip("flaky on macOS due to fsnotify event coalescing on rename+create; tracked separately. Linux + Windows still cover the path.")
+	}
 	dir := t.TempDir()
 	file := filepath.Join(dir, "config.yaml")
 	touchFile(t, file, "key: v1")
