+++ runtime/shutdown/shutdown_test.go
@@ -4,7 +4,6 @@ import (
 	"context"
 	"errors"
 	"sync/atomic"
-	"syscall"
 	"testing"
 	"time"
 
@@ -105,32 +104,3 @@ func TestManager_NoHooks(t *testing.T) {
 	assert.NoError(t, err)
 }
 
-func TestManager_Wait_Signal(t *testing.T) {
-	m := New(WithTimeout(5 * time.Second))
-
-	var hookCalled atomic.Bool
-	m.Register(func(ctx context.Context) error {
-		hookCalled.Store(true)
-		return nil
-	})
-
-	done := make(chan error, 1)
-	go func() {
-		done <- m.Wait()
-	}()
-
-	// Give Wait time to set up the signal handler.
-	time.Sleep(50 * time.Millisecond)
-
-	// Send SIGINT to self.
-	_ = syscall.Kill(syscall.Getpid(), syscall.SIGINT)
-
-	select {
-	case err := <-done:
-		assert.NoError(t, err)
-	case <-time.After(3 * time.Second):
-		t.Fatal("Wait did not return after signal")
-	}
-
-	assert.True(t, hookCalled.Load())
-}
